channel_requester: Auto-move clients to existing channels
A client can join the channel requester, which creates a channel for this user and moves the client
respectively. However, when the client leaves the channel and joins the channel requester channel
instead while the own channel still exists, the client stays in the channel requester channel. This
change ensures, that either a new channel gets created or that the client gets moved to the already
existing channel.